[Game] Implement laser beam collision.

+++ shared/src/action.rs
@@ -10,7 +10,7 @@ pub enum Action {
     FiredLaserBeam(u8, f32),
     // TODO use smaller values
     // TODO create helper for f32 <> u16
-    CreateLaserBeam(u8, f32, f32, f32, u8)
+    CreateLaserBeam(u8, f32, f32, f32, f32)
 }
 
 impl Action {
